requirements.txt
kagglehub==0.3.6
numpy==2.0.2
matplotlib==3.10.0
# tensorflow[and-cuda]==2.18.0 # Only use on Linux systems
tensorflow==2.18.0 # (if not using GPU) This does not have randomShear
opencv-python==4.10.0.84
asyncio==3.4.3
alive-progress==3.2.0
scikit-learn==1.6.1
PyQt6==6.8.0
colorlog==6.9.0
dotenv==0.9.9
django==5.1.7
djangorestframework==3.15.2
django-tailwind==4.0.1
django-browser-reload==1.18.0
pillow==11.1.0
psycopg2-binary==2.9.10
celery==5.3.1
redis==4.6.0
gevent==24.11.1